+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
 - **New:** Seed for server error logs (#45)
 - **New:** Seed for spam trigger log (#46)
 - **New:** Seed for spam cleaner log (#47)
+- **New:** Seed for email bounce log (#48)
 - **Fix:** Threads cannot be seeded (#36)
 
 ## 1.1.0 Alpha 3 (`1010013`)